+++ jobboard-core/src/main/java/com/advanceweb/afc/jb/employer/service/EmployerRegistration.java
@@ -46,7 +46,7 @@ public class EmployerRegistration implements ProfileRegistration {
 	 * 
 	 * @param profileId
 	 */
-	public boolean deleteProfile(long profileId){
+	public boolean deleteProfile(int profileId){
 		return false;
 	}
 
@@ -62,7 +62,7 @@ public class EmployerRegistration implements ProfileRegistration {
 	 * 
 	 * @param profileId
 	 */
-	public ProfileDTO viewProfile(long profileId){
+	public ProfileDTO viewProfile(int profileId){
 		return null;
 	}
